refactor: 코드 정리
- 패키지 생성과 클래스 이동
- RacingValidation, RacingValidationTest 클래스 생성
- RacingRule 클래스 내 메소드 일부를 RacingValidation 클래스로 이동
- RacingRuleTest 클래스 내 메소드 일부를 RacingValidationTest 클래스로 이동
- RacingConstant 클래스 생성
- RacingRule 클래스의 상수를 RacingConstant 클래스에서 관리하도록 변경
