+++ src/main/java/nextstep/utils/RacingRule.java
@@ -1,13 +1,13 @@
 package nextstep.utils;
 
-import racinggame.Message;
-import racinggame.RacingCar;
-import racinggame.RacingRuleException;
+import racinggame.common.constant.RacingConstant;
+import racinggame.common.enums.Message;
+import racinggame.common.validation.RacingValidation;
+import racinggame.domain.RacingCar;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.regex.Pattern;
 
 /**
  * @author Choi InJoo <rinjyu@naver.com>
@@ -15,65 +15,6 @@ import java.util.regex.Pattern;
  * @since 1.0
  */
 public class RacingRule {
-
-    public static final String RACING_CAR_DELIMITER = ",";
-
-    private static final String RACING_ADD_POINT = "-";
-
-    /**
-     * 데이터의 null or empty string 여부 확인
-     * @param data 데이터
-     * @return 데이터가 null 혹은 empty string 인지의 여부(true / false)
-     */
-    public static boolean isDataExists(String data) {
-        return !(data == null || data.trim().isEmpty());
-    }
-
-    /**
-     * 자동차명 내의 콤마(,) 포함 여부 확인
-     * @param racingCars 사용자가 입력한 자동차명
-     * @return 자동차명 내의 콤마(,)가 포함되어 있는 지의 여부(true / false)
-     */
-    public static boolean isRacingCarsNameContainsComma(String racingCars) {
-        return racingCars.contains(RACING_CAR_DELIMITER);
-    }
-
-    /**
-     * 자동차명의 글자수 확인
-     * @param racingCars 사용자가 입력한 자동차명
-     * @return 자동차명의 글자수가 1~5사이의 값인지의 여부(true / false)
-     */
-    public static boolean isEachRacingCarNameLength(String[] racingCars) {
-        for (String car : racingCars) {
-            if (!(car.length() > 0 && car.length() < 6)) return false;
-        }
-        return true;
-    }
-
-    /**
-     * 자동차명 유효성 확인
-     * @param racingCars 사용자가 입력한 자동차명
-     * @throws RacingRuleException 유효한 조건에 부합하지 않는 경우 발생
-     */
-    public static void isRacingCarsValidation(String racingCars) {
-        if (isDataExists(racingCars) && isRacingCarsNameContainsComma(racingCars) &&
-                isEachRacingCarNameLength(racingCars.split(RACING_CAR_DELIMITER))) {
-            return;
-        }
-        throw new RacingRuleException(Message.ERROR_INVALID_RACING_CAR_NAME.getMessage());
-    }
-
-    /**
-     * 횟수가 유효한 값인지의 여부 확인
-     * @param count 횟수
-     * @throws RacingRuleException 유효한 조건에 부합하지 않는 경우 발생
-     */
-    public static void isCountValidation(String count) {
-        if (Pattern.matches("^[0-9]*$", count) && Integer.parseInt(count) > 0) {
-            return;
-        }
-        throw new RacingRuleException(Message.ERROR_INVALID_NUMBER.getMessage());
-    }
 
     /**
      * 사용자가 입력한 자동차명을 List 형으로 변환
@@ -104,10 +45,10 @@ public class RacingRule {
      * @return 자동차의 현재 위치
      */
     public static String currentLocation(int count, String location) {
-        if (!isDataExists(location)) {
+        if (!RacingValidation.isDataExists(location)) {
             location = "";
         }
-        return RacingRule.isForward(count) ? (location + RACING_ADD_POINT) : location;
+        return RacingRule.isForward(count) ? (location + RacingConstant.RACING_ADD_POINT) : location;
     }
 
     /**
@@ -145,7 +86,7 @@ public class RacingRule {
      */
     public static void getRacingWinner(StringBuilder winner, String location, int maxDistance, String racingCarName) {
         if (location.length() == maxDistance) {
-            winner.append(isDataExists(winner.toString()) ? (RACING_CAR_DELIMITER + racingCarName) : racingCarName);
+            winner.append(RacingValidation.isDataExists(winner.toString()) ? (RacingConstant.RACING_CAR_DELIMITER + racingCarName) : racingCarName);
         }
     }
