+++ src/main/java/racinggame/common/constant/RacingConstant.java
@@ -1,0 +1,13 @@
+package racinggame.common.constant;
+
+/**
+ * @author Choi InJoo <rinjyu@naver.com>
+ * @version 1.0
+ * @since 1.0
+ */
+public class RacingConstant {
+
+    public static final String RACING_CAR_DELIMITER = ",";
+
+    public static final String RACING_ADD_POINT = "-";
+}
