+++ src/test/java/nextstep/utils/RacingRuleTest.java
@@ -2,10 +2,13 @@ package nextstep.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.*;
-import racinggame.Message;
-import racinggame.RacingCar;
-import racinggame.RacingRuleException;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.common.constant.RacingConstant;
+import racinggame.common.validation.RacingValidation;
+import racinggame.domain.RacingCar;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -17,64 +20,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 class RacingRuleTest {
-
-    private static final String RACING_CAR_DELIMITER = ",";
-
-    private static final String RACING_ADD_POINT = "-";
-
-    @ParameterizedTest
-    @DisplayName("데이터의 null or empty string 여부 확인")
-    @NullAndEmptySource
-    @ValueSource(strings = {"MyCar"})
-    void 데이터의_null_or_empty_string_여부_확인(String data) {
-        assertTrue(data == null || data.trim().isEmpty(), "입력 데이터 있음");
-    }
-
-    @ParameterizedTest
-    @DisplayName("자동차명 내의 콤마(,) 포함 여부 확인")
-    @ValueSource(strings = {"MyCar1", "MyCar2,MyCar3"})
-    void 자동차명_내의_콤마_포함_여부_확인(String racingCars) {
-        assertThat(racingCars).contains(RACING_CAR_DELIMITER);
-    }
-
-    @ParameterizedTest
-    @DisplayName("자동차명의 글자수 확인")
-    @ValueSource(strings = {"MyCar", "MyCar1234"})
-    void 자동차명의_글자수_확인(String racingCars) {
-        assertThat(racingCars.length()).isBetween(1, 5);
-    }
-
-    @ParameterizedTest
-    @DisplayName("자동차명 유효성 확인 : 유효한 값인 경우")
-    @ValueSource(strings = {"Car1,Car2"})
-    void 자동차명_유효성_확인_유효한_값인_경우(String racingCars) {
-        assertDoesNotThrow(() -> RacingRule.isRacingCarsValidation(racingCars));
-    }
-
-    @ParameterizedTest
-    @DisplayName("자동차명 유효성 확인 : 유효하지 않은 값인 경우")
-    @ValueSource(strings = {"MyCar,MyCar1", "MyCar1,MyCar", "MyCar1,MyCar2"})
-    void 자동차명_유효성_확인_유효하지_않은_값인_경우(String racingCars) {
-        RacingRuleException thrown = assertThrows(RacingRuleException.class,
-                () -> RacingRule.isRacingCarsValidation(racingCars));
-        assertThat(thrown.getMessage()).isEqualTo(Message.ERROR_INVALID_RACING_CAR_NAME.getMessage());
-    }
-
-    @ParameterizedTest
-    @DisplayName("횟수가 유효한 값인지의 여부 확인 : 유효한 값인 경우")
-    @ValueSource(strings = {"1", "2", "6", "8"})
-    void 횟수가_유효한_값인지의_여부_확인_유효한_값인_경우(String count) {
-        assertDoesNotThrow(() -> RacingRule.isCountValidation(count));
-    }
-
-    @ParameterizedTest
-    @DisplayName("횟수가 유효한 값인지의 여부 확인 : 유효하지 않은 값인 경우")
-    @ValueSource(strings = {"abc"})
-    void 횟수가_유효한_값인지의_여부_확인_유효하지_않은_값인_경우(String count) {
-        RacingRuleException thrown = assertThrows(RacingRuleException.class,
-                () -> RacingRule.isCountValidation(count));
-        assertThat(thrown.getMessage()).isEqualTo(Message.ERROR_INVALID_NUMBER.getMessage());
-    }
 
     @ParameterizedTest
     @DisplayName("사용자가 입력한 자동차명을 List 형으로 변환")
@@ -98,10 +43,10 @@ class RacingRuleTest {
     @DisplayName("각 자동차별 현재 위치 : 경주 시작")
     @CsvSource({"5,", "5,''"})
     void 각_자동차별_현재_위치_경주_시작(int count, String location) {
-        if (!RacingRule.isDataExists(location)) {
+        if (!RacingValidation.isDataExists(location)) {
             location = "";
         }
-        location = RacingRule.isForward(count) ? (location + RACING_ADD_POINT) : location;
+        location = RacingRule.isForward(count) ? (location + RacingConstant.RACING_ADD_POINT) : location;
         assertThat(location).isEqualTo("-");
     }
 
@@ -109,10 +54,10 @@ class RacingRuleTest {
     @DisplayName("각 자동차별 현재 위치 : 전진")
     @CsvSource({"6,'---'"})
     void 각_자동차별_현재_위치_전진(int count, String location) {
-        if (!RacingRule.isDataExists(location)) {
+        if (!RacingValidation.isDataExists(location)) {
             location = "";
         }
-        location = RacingRule.isForward(count) ? (location + RACING_ADD_POINT) : location;
+        location = RacingRule.isForward(count) ? (location + RacingConstant.RACING_ADD_POINT) : location;
         assertThat(location).isEqualTo("----");
     }
 
@@ -120,10 +65,10 @@ class RacingRuleTest {
     @DisplayName("각 자동차별 현재 위치 : 정지")
     @CsvSource({"3,'---'"})
     void 각_자동차별_현재_위치_정지(int count, String location) {
-        if (!RacingRule.isDataExists(location)) {
+        if (!RacingValidation.isDataExists(location)) {
             location = "";
         }
-        location = RacingRule.isForward(count) ? (location + RACING_ADD_POINT) : location;
+        location = RacingRule.isForward(count) ? (location + RacingConstant.RACING_ADD_POINT) : location;
         assertThat(location).isEqualTo("---");
     }
 
@@ -154,7 +99,7 @@ class RacingRuleTest {
     @MethodSource("getRacingCarWinner")
     void 자동차_경주_우승자_조회(StringBuilder winner, String location, int maxDistance, String racingCarName) {
         if (location.length() == maxDistance) {
-            winner.append(RacingRule.isDataExists(winner.toString()) ? (RACING_CAR_DELIMITER + racingCarName) : racingCarName);
+            winner.append(RacingValidation.isDataExists(winner.toString()) ? (RacingConstant.RACING_CAR_DELIMITER + racingCarName) : racingCarName);
         }
         assertAll(
                 () -> assertNotNull(winner.toString()),
@@ -173,7 +118,7 @@ class RacingRuleTest {
                 break;
             }
             if (racingCar.getLocation().length() == maxDistance) {
-                winner.append(RacingRule.isDataExists(winner.toString()) ? (RACING_CAR_DELIMITER + racingCar.getName()) : racingCar.getName());
+                winner.append(RacingValidation.isDataExists(winner.toString()) ? (RacingConstant.RACING_CAR_DELIMITER + racingCar.getName()) : racingCar.getName());
             }
         }
